Module Btree_on_file.Make1

Parameters

S : S

Signature

module M : sig ... end

module B : sig ... end

val make : store:(B.r, B.node) Mini_btree.Btree_impl_intf.store_ops -> alloc:(unit -> B.r Mini_btree.Btree_impl_intf.m) -> (B.k, B.v, B.r) Mini_btree.Btree_impl_intf.btree_ops

type header = {

mutable root : B.r;

mutable min_free_blk : int;

}

val bin_shape_header : Bin_prot.Shape.t

val bin_size_header : header Bin_prot.Size.sizer

val bin_write_header : header Bin_prot.Write.writer

val bin_writer_header : header Bin_prot.Type_class.writer

val __bin_read_header__ : (int -> header) Bin_prot.Read.reader

val bin_read_header : header Bin_prot.Read.reader

val bin_reader_header : header Bin_prot.Type_class.reader

val bin_header : header Bin_prot.Type_class.t

type cache_entry = {

node : B.node;

mutable dirty : bool;

}

type cache = (B.r, cache_entry) Base.Hashtbl.t

module R = Base.Int

val empty_cache : unit -> cache

type t = {

fn : string;

fd : Lwt_unix.file_descr;

header : header;

uncached_store : (B.r, B.node) Mini_btree.Btree_impl_intf.store_ops;

store_ops : (B.r, B.node) Mini_btree.Btree_impl_intf.store_ops;

btree_ops : (B.k, B.v, B.r) Mini_btree.Btree_impl_intf.btree_ops;

mutable closed : bool;

cache : cache;

}

val write_blk_bytes : Lwt_unix.file_descr -> int -> bytes -> unit Lwt.t

val write_blk_ba : Lwt_unix.file_descr -> int -> Bigstringaf.t -> unit Lwt.t

val read_blk_bytes : Lwt_unix.file_descr -> int -> bytes Lwt.t

val read_blk_ba : Lwt_unix.file_descr -> int -> Bigstringaf.t Lwt.t

module H : sig ... end

module From_fd : functor (S : sig ... end) -> sig ... end

module Cached_store : sig ... end

This is too simple: at the moment it caches every read and write, and only goes to disk on a flush. OK for a demo but should be replaced with an LRU or similar.

module From_store : functor (S : sig ... end) -> sig ... end

val create : fn:string -> t Lwt.t

val open_ : fn:string -> t Lwt.t

val find : t -> B.k -> B.v option Mini_btree.Btree_impl_intf.m

val insert : t -> B.k -> B.v -> unit Lwt.t

val insert_many : t -> (B.k * B.v) list -> (B.k * B.v) list Lwt.t

val delete : t -> B.k -> unit Mini_btree.Btree_impl_intf.m

val close : t -> unit Lwt.t