Module Mini_btree__.Mmap_intf

type bigstring = (char, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

module type S = sig ... end

To write to/from the mmap, use the sub function to get the subarray of interest. NOTE that because the map may grow when asking for a sub which extends beyond the current length, a remap may occur. Thus, subs should not be reused or stored, since they may refer to an out of date slice for an old mapping.