Module Mini_btree__Example

An example instance using bin_prot for marshalling, and a backing file to store the nodes of the B-tree

val blk_sz : int

type buf = bytes

NOTE this is what lwt prefers, but bin_prot prefers bigarray :( Bin_prot.Common has some useful blitting functions... also in Bigstringaf

val ba_to_bytes : Bigstringaf.t -> Stdlib.Bytes.t

val bytes_to_ba : Stdlib.Bytes.t -> Bigstringaf.t

type 'r blk_dev_ops = {

read : 'r -> buf Mini_btree.Btree_impl_intf.m;

write : 'r -> buf -> unit Mini_btree.Btree_impl_intf.m;

}

Provided by the underlying block device, via lwt pread/pwrite on a file for example

module Marshalling_with_bin_prot : functor (S : sig ... end) -> sig ... end

module Int_int_int : sig ... end

module Marshalling_int_int : sig ... end

module Private_btree_on_file : sig ... end

Simple example, where we store the B-tree blocks in a file

module Btree_on_file : sig ... end