Module Private_btree_on_file.Cached_store

This is too simple: at the moment it caches every read and write, and only goes to disk on a flush. OK for a demo but should be replaced with an LRU or similar.

module Hashtbl = Base.Hashtbl

val cache_store : ($E.r, cache_entry) Hashtbl.t -> ($E.r, $E.node) Mini_btree.Btree_impl_intf.store_ops -> ($E.r Hashtbl.key, $E.node) Mini_btree.Btree_impl_intf.store_ops

val flush_cache : ('a, cache_entry) Hashtbl.t -> ('a Hashtbl.key, $E.node) Mini_btree.Btree_impl_intf.store_ops -> unit Lwt.t