Module Mini_btree__Mmap_private

Additional support for mmap; we want to have: 1) growable files; 2) ability to track the max used part of the file

module Bigarray = Stdlib.Bigarray

type buffer = Mini_btree__.Mmap_intf.bigstring

type t = {

fd : Unix.file_descr;

mutable buf : buffer;

}

module type S = sig ... end

What we need to use mmap; for large file sizes, we need to be on a 64 bit machine

module Make_1 : functor (S : S) -> sig ... end