Module Mini_btree__Simple_cache

A simple cache, using a hashtbl and a dirty flag.

Users are expected to insert elements and take care when elements are expunged.

module type T = sig ... end

What we get from creating a cache

module Make : functor (S : sig ... end) -> T with type k = S.k and type v = S.v

module Make_map_cache : functor (S : sig ... end) -> sig ... end

Cache for find, insert, delete operations